requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.102
botocore==1.34.102
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.5
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.56
marshmallow==3.21.2
momento==1.20.1
momento-wire-types==0.106.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.28.0
orjson==3.10.3
packaging==23.2
protobuf==4.25.3
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.10
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.27.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.6.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
